requirements.txt
Flask==2.3.2
qrcode==7.3
Pillow==10.0.0
numpy==1.26.0
